Capture Enter while a search is going with preliminary results suppressed.
Suppressing preliminary results is used when the expectation is that the user is typing input that will result in a single match, e.g. a Work Item ID. This is used as a heuristic to capture the Enter key being pressed. If the search resulted in a single match, that match is then executed.