Implemented proper algorithm to resolve plugin entry points.
I made an assumption that a package would contain a DLL with the same name as the package. This assumption broke when I changed all package IDs to start with TQLApp. The new implementation loads all assemblies of an installed plugin into a separate app domain (requires an alternative implementation on .NET Core) and looks for plugins in all assemblies. This is written to a manifest file for performance.